Move from CGI::XMLPost to CGI
It would be convenient to use some of the features available in the
CGI module, but CGI::XMLPost and CGI cannot (easily) co-exist.

As an added bonus, this means that we have one less external
non-core Perl module dependency, and might make future persistency
under mod_perl or Plack easier.

Signed-off-by: Jeff Godin <jgodin@tadl.org>