Update pickup lib at ItemShipped time
The ItemRequested NCIP message does not include information on where
the user wishes to pick up their hold.

Because of this, all holds are created with an initial hold pickup
location from the ini file.

Pickup library information is included in the ItemShipped message.

When we receive the ItemShipped message, we retrieve the hold and
update its pickup library based on a mapping file which maps the
address data in the NCIP message to a local org unit id.

Signed-off-by: Jeff Godin <jgodin@tadl.org>