Agency user from conf file
Would this work in a general sense? Sorry I'm not in a position to test directly, but would like to help get an iNCIPit fork fully aligned with this main repository. This is perhaps a stop-gap from fulfilling the requirements mentioned in the TODO, and may not be appropriate in every case? But if it works it's one of two points of divergence that can be eliminated.